+++ kryptokrona-http/src/main/kotlin/org/kryptokrona/sdk/http/model/response/queryblocks/QueryBlocks.kt
@@ -36,32 +36,32 @@ import org.kryptokrona.sdk.http.model.response.transaction.TransactionPrefix
 
 @Serializable
 data class QueryBlocks(
-    @SerialName("current_height") val currentHeight: Int,
-    @SerialName("full_offset") val fullOffset: Int,
+    @SerialName("current_height") val currentHeight: Long,
+    @SerialName("full_offset") val fullOffset: Long,
     val items: List<QueryBlocksItems>,
-    @SerialName("start_height") val startHeight: Int,
+    @SerialName("start_height") val startHeight: Long,
     val status: String
 )
 
 @Serializable
 data class QueryBlocksItems(
-    @SerialName("blockFullInfo.block") val block: List<Int>,
+    @SerialName("blockFullInfo.block") val block: List<Long>,
     @SerialName("blockFullInfo.block_id") val blockId: String,
     @SerialName("txs") val transactions: List<String>
 )
 
 @Serializable
 data class QueryBlocksLite(
-    val currentHeight: Int,
-    val fullOffset: Int,
+    val currentHeight: Long,
+    val fullOffset: Long,
     val items: List<QueryBlocksLiteItems>,
-    val startHeight: Int,
+    val startHeight: Long,
     val status: String
 )
 
 @Serializable
 data class QueryBlocksLiteItems(
-    @SerialName("blockShortInfo.block") val block: List<Int>,
+    @SerialName("blockShortInfo.block") val block: List<Long>,
     @SerialName("blockShortInfo.blockId") val blockId: String,
     @SerialName("blockShortInfo.txPrefixes") val transactionPrefixes: List<TransactionPrefix>
 )
